services/doc-pipeline/pyproject.toml
[project]
name = "bazan-doc-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.3.0",
    "pdfplumber>=0.11.0",
    "httpx>=0.27.0",
    "ollama>=0.2.0",
]

[project.optional-dependencies]
ocr = ["surya-ocr>=0.4.0", "pymupdf>=1.24.0", "pillow>=10.0.0"]

[tool.ruff]
line-length = 100
target-version = "py312"
